requirements.txt
torch
torchvision
numpy
tqdm
Pillow
scikit-image
opencv-python
